restructure project layout
